Refactor get_first_part into main package
